% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_height.R
\name{add_figure_height}
\alias{add_figure_height}
\title{Add figure_height column to data frame}
\usage{
add_figure_height(df)
}
\arguments{
\item{df}{}
}
\value{
Data frame with additional column
}
\description{
Add figure_height column to data frame
}
\examples{
add_figure_height(df)
}
