% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_rates.R
\name{add_response_rate}
\alias{add_response_rate}
\title{Adds summary row with overall response rates to data frame}
\usage{
add_response_rate(df, label)
}
\arguments{
\item{df}{Data frame obtained by \code{extract_ft_data_from_gt}}

\item{label}{Label for that survey type}
}
\value{
Data frame
}
\description{
Adds summary row with overall response rates to data frame
}
