% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_rates.R
\name{extract_ft_data_from_gt}
\alias{extract_ft_data_from_gt}
\title{Converts gt object to ft object and extracts data frame}
\usage{
extract_ft_data_from_gt(gt, label, col_index, into_list)
}
\arguments{
\item{gt}{Gt object}

\item{label}{Label for that survey type}

\item{col_index}{Integer vector of gt stat columns to split}

\item{into_list}{List of gt stat column names for each split}
}
\value{
Data frame
}
\description{
Converts gt object to ft object and extracts data frame
}
