% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_template.R
\name{get_chunk_parameters}
\alias{get_chunk_parameters}
\title{Get parameters for creating the template chunks plotting all figures}
\usage{
get_chunk_parameters(df, function_call = "RUBer::plot_figure")
}
\arguments{
\item{df}{Data frame with columns figure_count}
}
\value{
Data frame with columns chunk_label, function_parameters and
function_call
}
\description{
Get parameters for creating the template chunks plotting all figures
}
\examples{
get_chunk_parameters(RUBer::df_example)
}
