% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funded_projects.R
\name{get_funded_projects}
\alias{get_funded_projects}
\title{Get data frame of funded projects}
\usage{
get_funded_projects(path)
}
\arguments{
\item{path}{Character, path to csv file}
}
\value{
Data frame
}
\description{
Get data frame of funded projects
}
