% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_height.R
\name{get_max_figure_height}
\alias{get_max_figure_height}
\title{Get max figure height}
\usage{
get_max_figure_height(
  figure_caption,
  heading,
  is_heading,
  subheading,
  is_subheading
)
}
\arguments{
\item{figure_caption}{Text}

\item{heading}{Text}

\item{is_heading}{Boolean}

\item{subheading}{Text}

\item{is_subheading}{Boolean}
}
\value{
Max figure height in inches
}
\description{
Get max figure height
}
\examples{
get_max_figure_height(
  figure_caption = "Beurteilung der Studienangebote und -bedingungen - Struktur des Studiums (Bachelor)",
  heading = NA_character_,
  is_heading = FALSE,
  subheading = "Beurteilung von Studium und Lehre Beurteilung von Studium und Lehre asdfs",
  is_subheading = TRUE
)
}
