% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_df.R
\name{get_report_df}
\alias{get_report_df}
\title{Cleans up the report data}
\usage{
get_report_df(df)
}
\arguments{
\item{df}{Data frame obtained by load_raw_reports}
}
\value{
Data frame
}
\description{
Cleans up the report data
}
