% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_rates.R
\name{get_response_rate_data}
\alias{get_response_rate_data}
\title{Get data on response rates for a given \code{survey_type_id}}
\usage{
get_response_rate_data(
  path,
  survey_type_id,
  label,
  col_index = c(2L, 4L),
  into_list = list(c("koepfe_rub", "koepfe_rub_perc"), c("koepfe_bef",
    "koepfe_bef_perc"))
)
}
\arguments{
\item{path}{Path to csv file with raw data}

\item{survey_type_id}{One of \code{c("AB", "EB", "VB")}}

\item{label}{Label for that survey type}

\item{col_index}{Integer vector of gt stat columns to split}

\item{into_list}{List of gt stat column names for each split}
}
\value{
Data frame
}
\description{
Get data on response rates for a given \code{survey_type_id}
}
