% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stg_df.R
\name{get_stg_df}
\alias{get_stg_df}
\title{Filters data frame based on report_nr and creates factors studiengang and
studienfachzaehler.}
\usage{
get_stg_df(df, report_nr)
}
\arguments{
\item{df}{Data frame obtained by load_raw_studienfaelle}

\item{report_nr}{Report_nr to filter the data}
}
\value{
Filtered data frame with factor columns
}
\description{
Filters data frame based on report_nr and creates factors studiengang and
studienfachzaehler.
}
