% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_height.R
\name{get_variable_height}
\alias{get_variable_height}
\title{Add figure_height to data frame}
\usage{
get_variable_height(df)
}
\arguments{
\item{df}{}
}
\value{
Data frame with additional column figure_height
}
\description{
Add figure_height to data frame
}
