% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw_data.R
\name{load_raw_excluded_programs}
\alias{load_raw_excluded_programs}
\title{Load csv file on excluded programs}
\usage{
load_raw_excluded_programs(path)
}
\arguments{
\item{path}{Character, path to csv file}
}
\value{
Data frame
}
\description{
Load csv file on excluded programs
}
