% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw_data.R
\name{load_raw_funded_projects}
\alias{load_raw_funded_projects}
\title{Load csv file on funded projects}
\usage{
load_raw_funded_projects(path)
}
\arguments{
\item{path}{Character, path to csv file}
}
\value{
Data frame
}
\description{
Load csv file on funded projects
}
