% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw_data.R
\name{load_raw_included_programs}
\alias{load_raw_included_programs}
\title{Load csv file on included programs}
\usage{
load_raw_included_programs(path)
}
\arguments{
\item{path}{Character, path to csv file}
}
\value{
Data frame
}
\description{
Load csv file on included programs
}
