% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw_data.R
\name{load_raw_metrics}
\alias{load_raw_metrics}
\title{Load csv file on metrics}
\usage{
load_raw_metrics(path)
}
\arguments{
\item{path}{Character, path to csv file}
}
\value{
Data frame
}
\description{
Load csv file on metrics
}
