% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw_data.R
\name{load_raw_reports}
\alias{load_raw_reports}
\title{Loads csv file with information on reports}
\usage{
load_raw_reports(path)
}
\arguments{
\item{path}{Path to csv file with raw data}
}
\value{
Data frame
}
\description{
Loads csv file with information on reports
}
