% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw_data.R
\name{load_raw_response_rates}
\alias{load_raw_response_rates}
\title{Loads csv file with information on response rates}
\usage{
load_raw_response_rates(path)
}
\arguments{
\item{path}{Path to csv file with raw data}
}
\value{
Data frame
}
\description{
Loads csv file with information on response rates
}
