% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw_data.R
\name{load_raw_studienfaelle}
\alias{load_raw_studienfaelle}
\title{Loads csv file with information on cases}
\usage{
load_raw_studienfaelle(path)
}
\arguments{
\item{path}{Path to csv file with raw data}
}
\value{
Data frame
}
\description{
Loads csv file with information on cases
}
