% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_umlauts.R
\name{replace_umlauts}
\alias{replace_umlauts}
\title{Replace German umlauts with ASCII escape sequence for unicode characters}
\usage{
replace_umlauts(path)
}
\arguments{
\item{path}{Character, path to file}
}
\value{
Side effects
}
\description{
Replace German umlauts with ASCII escape sequence for unicode characters
}
