% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_rates.R
\name{response_rate_gt}
\alias{response_rate_gt}
\title{Creates gt object of response rates for a given \code{survey_type_id}}
\usage{
response_rate_gt(df, survey_type_id)
}
\arguments{
\item{df}{Data frame obtained by \code{wrangle_response_rates}}

\item{survey_type_id}{One of \code{c("AB", "EB", "VB")}}
}
\value{
Gt object
}
\description{
Creates gt object of response rates for a given \code{survey_type_id}
}
