% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_figure_df.R
\name{set_factors}
\alias{set_factors}
\title{Type conversions for the 2022 data}
\usage{
set_factors(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with the factors y, fill, fill_label and group
}
\description{
Type conversions for the 2022 data
}
\examples{
\dontrun{
set_factors(df)
}
}
