% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_figures.R
\name{split_figures}
\alias{split_figures}
\title{Splits up big figures}
\usage{
split_figures(df)
}
\arguments{
\item{df}{}
}
\value{
Df
}
\description{
Splits up big figures
}
\examples{
split_figures(df)
}
