% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_templates.R
\name{tpl_get_figure_df}
\alias{tpl_get_figure_df}
\title{Returns a character vector for the code chunk to retrieve the figure data
frame using \code{Datenreport2022::get_figure_df}}
\usage{
tpl_get_figure_df()
}
\value{
Character vector
}
\description{
Returns a character vector for the code chunk to retrieve the figure data
frame using \code{Datenreport2022::get_figure_df}
}
\examples{
writeLines(tpl_get_figure_df())
}
