% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_templates.R
\name{tpl_plot_figure}
\alias{tpl_plot_figure}
\title{Returns a character vector for the code chunk to plot a figure}
\usage{
tpl_plot_figure(
  chunk_label,
  tab_placeholder = "PLACEHOLDER_TAB",
  figure_caption,
  figure_height,
  suppress_warnings = TRUE,
  function_call = "RUBer::plot_figure"
)
}
\arguments{
\item{chunk_label}{Character, chunk label}

\item{tab_placeholder}{Character, pandoc does not support the insertion of
tabs, which is why a placeholder text is needed that
will get replaced with tabs in post-processing,
defaults to "PLACEHOLDER_TAB"}

\item{figure_caption}{Character, figure caption}

\item{figure_height}{Numeric, figure height in inches}

\item{suppress_warnings}{Boolean, whether to suppress the warnings generated by this
code chunk, defaults to \code{TRUE}.}

\item{function_call}{Character, defaults to \code{RUBer::plot_figure}}
}
\value{
List of character vectors with chunk texts
}
\description{
Returns a character vector for the code chunk to plot a figure
}
\examples{
writeLines(tpl_plot_figure())
}
