% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_templates.R
\name{tpl_subheading}
\alias{tpl_subheading}
\title{Returns a character vector for the code chunk to insert a subheading}
\usage{
tpl_subheading(chunk_label, subheading)
}
\arguments{
\item{chunk_label}{Character, chunk label}

\item{subheading}{Character, subheading text}
}
\value{
List of character vectors with chunk texts
}
\description{
Returns a character vector for the code chunk to insert a subheading
}
\examples{
writeLines(tpl_subheading())
}
