% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_rates.R
\name{wrangle_response_rates}
\alias{wrangle_response_rates}
\title{Transforms response rates to long format, creates factors}
\usage{
wrangle_response_rates(df)
}
\arguments{
\item{df}{Data frame obtained by \code{load_raw_response_rates}}
}
\value{
Data frame
}
\description{
Transforms response rates to long format, creates factors
}
