% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funded_projects.R
\name{get_funded_projects}
\alias{get_funded_projects}
\title{Get data frame of funded projects}
\usage{
get_funded_projects(df, report_nr = NULL)
}
\arguments{
\item{df}{Data frame retrieved by \code{load_raw_funded_projects}}

\item{report_nr}{Optional integer, result set will be filtered to these report numbers}
}
\value{
Data frame
}
\description{
Get data frame of funded projects
}
