% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_templates.R
\name{tpl_heading}
\alias{tpl_heading}
\title{Returns a character vector for the code chunk to insert a heading}
\usage{
tpl_heading()
}
\value{
Character vector
}
\description{
Returns a character vector for the code chunk to insert a heading
}
\examples{
writeLines(tpl_heading())
}
