% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_templates.R
\name{tpl_plot_figure}
\alias{tpl_plot_figure}
\title{Returns a character vector for the code chunk to plot a figure using
\code{RUBer::plot_figure}}
\usage{
tpl_plot_figure(tab_placeholder = "PLACEHOLDER_TAB", suppress_warnings = TRUE)
}
\arguments{
\item{tab_placeholder}{Character, pandoc does not support the insertion of
tabs, which is why a placeholder text is needed that
will get replaced with tabs in post-processing,
defaults to "PLACEHOLDER_TAB"}

\item{warning}{Boolean, whether to suppress the warnings generated by this
code chunk, defaults to TRUE.}
}
\value{
Character vector
}
\description{
Returns a character vector for the code chunk to plot a figure using
\code{RUBer::plot_figure}
}
\examples{
writeLines(tpl_plot_figure())
}
