% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_templates.R
\name{tpl_subheading}
\alias{tpl_subheading}
\title{Returns a character vector for the code chunk to insert a subheading}
\usage{
tpl_subheading()
}
\value{
Character vector
}
\description{
Returns a character vector for the code chunk to insert a subheading
}
\examples{
writeLines(tpl_subheading())
}
